refactor: give the generic composite table no secondary indexes
The shared composite table now comes in two variants: an index-free one under
the original name, and an index-bearing one carrying the same LSIs and GSIs
under their existing names, since a number of tests pin error messages that
name them. Tests depending on an index move to files of their own, so a file's
table declaration says plainly whether it needs an index.
