+++ results/tag-manifest.json
@@ -42,6 +42,14 @@
         "data-plane"
       ]
     },
+    "tests/tier1/batchWriteItem/indexKeyValidation.test.ts": {
+      "BatchWriteItem — index key validation": [
+        "batch",
+        "data-plane",
+        "negative-path",
+        "lsi"
+      ]
+    },
     "tests/tier1/batchWriteItem/validation.test.ts": {
       "BatchWriteItem — validation": [
         "batch",
@@ -142,6 +150,14 @@
         "describe-table",
         "control-plane",
         "negative-path"
+      ]
+    },
+    "tests/tier1/describeTable/indexes.test.ts": {
+      "DescribeTable — secondary indexes": [
+        "describe-table",
+        "control-plane",
+        "gsi",
+        "lsi"
       ]
     },
     "tests/tier1/getItem/basic.test.ts": {
@@ -249,10 +265,19 @@
         "data-plane"
       ]
     },
+    "tests/tier1/putItem/indexKeyValidation.test.ts": {
+      "PutItem — index key validation": [
+        "put-item",
+        "data-plane",
+        "negative-path",
+        "lsi"
+      ]
+    },
     "tests/tier1/putItem/itemCollectionMetrics.test.ts": {
       "ReturnItemCollectionMetrics": [
         "put-item",
-        "data-plane"
+        "data-plane",
+        "lsi"
       ]
     },
     "tests/tier1/putItem/numberFormat.test.ts": {
@@ -332,6 +357,12 @@
         "query",
         "data-plane",
         "gsi"
+      ],
+      "Query — GSI ExclusiveStartKey validation": [
+        "query",
+        "data-plane",
+        "gsi",
+        "negative-path"
       ]
     },
     "tests/tier1/query/gsiHashOnlyPagination.test.ts": {
@@ -386,10 +417,6 @@
         "query",
         "data-plane"
       ],
-      "Query — Select modes on indexes": [
-        "query",
-        "data-plane"
-      ],
       "Query — Select SPECIFIC_ATTRIBUTES": [
         "query",
         "data-plane"
@@ -398,6 +425,14 @@
         "query",
         "data-plane",
         "negative-path"
+      ]
+    },
+    "tests/tier1/query/selectOnIndexes.test.ts": {
+      "Query — Select modes on indexes": [
+        "query",
+        "data-plane",
+        "gsi",
+        "lsi"
       ]
     },
     "tests/tier1/query/validation.test.ts": {
@@ -463,6 +498,11 @@
     },
     "tests/tier1/scan/gsi.test.ts": {
       "Scan — GSI": [
+        "scan",
+        "data-plane",
+        "gsi"
+      ],
+      "Scan — GSI FilterExpression parens": [
         "scan",
         "data-plane",
         "gsi"
@@ -740,6 +780,14 @@
         "data-plane"
       ]
     },
+    "tests/tier2/transactions/transactWriteIndexKeys.test.ts": {
+      "TransactWriteItems — index key validation": [
+        "transactions",
+        "data-plane",
+        "negative-path",
+        "lsi"
+      ]
+    },
     "tests/tier2/ttl/basic.test.ts": {
       "TTL — basic": [
         "ttl",
@@ -827,6 +875,40 @@
         "negative-path"
       ]
     },
+    "tests/tier3/error-messages/indexKeyValues.test.ts": {
+      "BatchWriteItem — index key error messages": [
+        "batch",
+        "data-plane",
+        "negative-path",
+        "gsi"
+      ],
+      "PutItem — index key error messages": [
+        "put-item",
+        "data-plane",
+        "negative-path",
+        "gsi"
+      ],
+      "UpdateItem — index key error messages": [
+        "update-item",
+        "data-plane",
+        "negative-path",
+        "gsi"
+      ]
+    },
+    "tests/tier3/error-messages/indexReads.test.ts": {
+      "Query — index error messages": [
+        "query",
+        "data-plane",
+        "negative-path",
+        "gsi"
+      ],
+      "Scan — index error messages": [
+        "scan",
+        "data-plane",
+        "negative-path",
+        "gsi"
+      ]
+    },
     "tests/tier3/error-messages/partiql.test.ts": {
       "PartiQL — exact error messages": [
         "partiql",
@@ -869,6 +951,14 @@
         "transactions",
         "data-plane",
         "negative-path"
+      ]
+    },
+    "tests/tier3/error-messages/transactIndexKeys.test.ts": {
+      "TransactWriteItems — index key error messages": [
+        "transactions",
+        "data-plane",
+        "negative-path",
+        "gsi"
       ]
     },
     "tests/tier3/error-messages/transactWriteItems.test.ts": {
