+++ src/setup.ts
@@ -7,17 +7,13 @@ import {
 } from './indeterminate-sink.js'
 import { clearObservedMarker } from './observation-sink.js'
 
-// Provision the shared tables the selected test files asked for.
+// Provision the shared tables the selected test files declared.
 //
-// vitest 4 runs setupFiles' beforeAll for every test file (vitest 3's singleFork
-// ran it once), and each file's declarations only register when that file is
-// imported, so this hook has to run per file to create what the newest file
-// added. Both steps are idempotent across those runs, and each is guarded in
-// src/helpers.ts for a different reason: tables are memoised by name, so a def
-// forty files declare is created once rather than deleted and recreated ~100
-// times a run; the sweep is memoised separately so it stays one pass at run
-// start and never deletes tables a later file is still using. Final teardown
-// runs once in src/global-teardown.ts.
+// vitest 4 runs this beforeAll per test file, and a file's declarations only
+// register once it is imported, so the hook must run per file to create what
+// the newest one added. Both steps are guarded in src/helpers.ts: tables are
+// memoised by name, the sweep separately so it stays one pass at run start.
+// Final teardown runs once in src/global-teardown.ts.
 beforeAll(async () => {
   try {
     await cleanupAllTablesOnce()
