+++ tests/tier1/batchWriteItem/validation.test.ts
@@ -2,12 +2,11 @@ import { BatchWriteItemCommand } from '@aws-sdk/client-dynamodb'
 import { ddb } from '../../../src/client.js'
 import {
   hashTableDef,
-  compositeTableDef,
   expectDynamoError,
   declareTables,
 } from '../../../src/helpers.js'
 
-declareTables(hashTableDef, compositeTableDef)
+declareTables(hashTableDef)
 
 describe('BatchWriteItem — validation', { tags: ['batch', 'data-plane', 'negative-path'] }, () => {
   it('rejects more than 25 items', async () => {
@@ -90,11 +89,11 @@ describe('BatchWriteItem — validation', { tags: ['batch', 'data-plane', 'negat
     }
   })
 
-  // A key attribute (table or index) whose value has the wrong type, is
-  // non-scalar, or is an empty string is rejected. BatchWriteItem validates the
-  // item up front, so every variant is a top-level ValidationException for the
-  // whole request - there is no cancellation path here, unlike TransactWriteItems.
-  // Exact strings are pinned in tests/tier3/error-messages/batchWriteItem.test.ts.
+  // A table key attribute whose value has the wrong type, is non-scalar, or is
+  // an empty string is rejected. BatchWriteItem validates the item up front, so
+  // every variant is a top-level ValidationException for the whole request -
+  // there is no cancellation path here, unlike TransactWriteItems. Exact
+  // strings are pinned in tests/tier3/error-messages/batchWriteItem.test.ts.
   it('rejects a wrong-typed table key value', async () => {
     await expectDynamoError(
       () => ddb.send(
@@ -140,66 +139,6 @@ describe('BatchWriteItem — validation', { tags: ['batch', 'data-plane', 'negat
       ),
       'ValidationException',
       /empty string value/i,
-    )
-  })
-
-  it('rejects a wrong-typed index key value', async () => {
-    await expectDynamoError(
-      () => ddb.send(
-        new BatchWriteItemCommand({
-          RequestItems: {
-            [compositeTableDef.name]: [
-              {
-                PutRequest: {
-                  Item: { pk: { S: 'bw-idx-type' }, sk: { S: 'a' }, lsi1sk: { N: '5' } },
-                },
-              },
-            ],
-          },
-        }),
-      ),
-      'ValidationException',
-      /Type mismatch for Index Key/,
-    )
-  })
-
-  it('rejects a non-scalar index key value', async () => {
-    await expectDynamoError(
-      () => ddb.send(
-        new BatchWriteItemCommand({
-          RequestItems: {
-            [compositeTableDef.name]: [
-              {
-                PutRequest: {
-                  Item: { pk: { S: 'bw-idx-nonscalar' }, sk: { S: 'b' }, lsi1sk: { L: [{ S: 'x' }] } },
-                },
-              },
-            ],
-          },
-        }),
-      ),
-      'ValidationException',
-      /Type mismatch for Index Key/,
-    )
-  })
-
-  it('rejects an empty-string index key value', async () => {
-    await expectDynamoError(
-      () => ddb.send(
-        new BatchWriteItemCommand({
-          RequestItems: {
-            [compositeTableDef.name]: [
-              {
-                PutRequest: {
-                  Item: { pk: { S: 'bw-idx-empty' }, sk: { S: 'c' }, lsi1sk: { S: '' } },
-                },
-              },
-            ],
-          },
-        }),
-      ),
-      'ValidationException',
-      /secondary index key/i,
     )
   })
 })
