+++ tests/tier1/describeTable/basic.test.ts
@@ -2,12 +2,11 @@ import { DescribeTableCommand } from '@aws-sdk/client-dynamodb'
 import { ddb } from '../../../src/client.js'
 import {
   hashTableDef,
-  compositeTableDef,
   expectDynamoError,
   declareTables,
 } from '../../../src/helpers.js'
 
-declareTables(hashTableDef, compositeTableDef)
+declareTables(hashTableDef)
 
 describe('DescribeTable — basic', { tags: ['describe-table', 'control-plane'] }, () => {
   it('returns table metadata for a hash-only table', async () => {
@@ -28,32 +27,6 @@ describe('DescribeTable — basic', { tags: ['describe-table', 'control-plane'] 
     expect(table.CreationDateTime).toBeDefined()
     expect(table.ItemCount).toBeDefined()
     expect(table.TableSizeBytes).toBeDefined()
-  })
-
-  it('returns table metadata for a composite table with indexes', async () => {
-    const result = await ddb.send(
-      new DescribeTableCommand({ TableName: compositeTableDef.name }),
-    )
-    const table = result.Table!
-
-    expect(table.TableName).toBe(compositeTableDef.name)
-    expect(table.KeySchema).toHaveLength(2)
-
-    // LSIs
-    expect(table.LocalSecondaryIndexes).toBeDefined()
-    expect(table.LocalSecondaryIndexes).toHaveLength(2)
-
-    // GSIs
-    expect(table.GlobalSecondaryIndexes).toBeDefined()
-    expect(table.GlobalSecondaryIndexes).toHaveLength(2)
-
-    // Each GSI should have IndexStatus
-    for (const gsi of table.GlobalSecondaryIndexes!) {
-      expect(gsi.IndexName).toBeDefined()
-      expect(gsi.IndexStatus).toBe('ACTIVE')
-      expect(gsi.KeySchema).toBeDefined()
-      expect(gsi.Projection).toBeDefined()
-    }
   })
 })
