+++ tests/tier1/query/lsi.test.ts
@@ -5,12 +5,12 @@ import {
 } from '@aws-sdk/client-dynamodb'
 import { ddb } from '../../../src/client.js'
 import {
-  compositeTableDef,
+  compositeIndexedTableDef,
   cleanupItems,
   declareTables,
 } from '../../../src/helpers.js'
 
-declareTables(compositeTableDef)
+declareTables(compositeIndexedTableDef)
 
 describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   const items = [
@@ -48,7 +48,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     await Promise.all(
       items.map((item) =>
         ddb.send(
-          new PutItemCommand({ TableName: compositeTableDef.name, Item: item }),
+          new PutItemCommand({ TableName: compositeIndexedTableDef.name, Item: item }),
         ),
       ),
     )
@@ -56,7 +56,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
 
   afterAll(async () => {
     await cleanupItems(
-      compositeTableDef.name,
+      compositeIndexedTableDef.name,
       items.map((item) => ({ pk: item.pk, sk: item.sk })),
     )
   })
@@ -64,7 +64,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('queries LSI with equality on sort key', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk AND lsi1sk = :sk',
         ExpressionAttributeValues: {
@@ -82,7 +82,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('queries LSI with begins_with on sort key', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk AND begins_with(lsi1sk, :prefix)',
         ExpressionAttributeValues: {
@@ -102,7 +102,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('queries LSI with BETWEEN on sort key', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi2',
         KeyConditionExpression: 'pk = :pk AND lsi2sk BETWEEN :lo AND :hi',
         ExpressionAttributeValues: {
@@ -124,7 +124,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('supports ConsistentRead on LSI (unlike GSI)', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk',
         ExpressionAttributeValues: { ':pk': { S: 'lsi-q-1' } },
@@ -138,7 +138,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('ALL projection returns all base table attributes', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk AND lsi1sk = :sk',
         ExpressionAttributeValues: {
@@ -163,7 +163,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     // lsi2 has projectionType INCLUDE with nonKeyAttributes ['lsi1sk']
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi2',
         KeyConditionExpression: 'pk = :pk AND lsi2sk = :sk',
         ExpressionAttributeValues: {
@@ -189,7 +189,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('returns items only from the queried partition', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk',
         ExpressionAttributeValues: { ':pk': { S: 'lsi-q-1' } },
@@ -206,7 +206,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
   it('returns empty results for non-existent partition key on LSI', async () => {
     const result = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :pk',
         ExpressionAttributeValues: { ':pk': { S: 'does-not-exist' } },
@@ -223,7 +223,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     const noLsiSkItem = { pk: { S: 'lsi-sparse-q' }, sk: { S: 'x' } }
     await ddb.send(
       new PutItemCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         Item: noLsiSkItem,
       }),
     )
@@ -231,7 +231,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     // Absent from the LSI...
     const lsiResult = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         IndexName: 'lsi1',
         KeyConditionExpression: 'pk = :p',
         ExpressionAttributeValues: { ':p': { S: 'lsi-sparse-q' } },
@@ -243,7 +243,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     // ...but present in the base table.
     const baseResult = await ddb.send(
       new QueryCommand({
-        TableName: compositeTableDef.name,
+        TableName: compositeIndexedTableDef.name,
         KeyConditionExpression: 'pk = :p',
         ExpressionAttributeValues: { ':p': { S: 'lsi-sparse-q' } },
         ConsistentRead: true,
@@ -251,7 +251,7 @@ describe('Query — LSI', { tags: ['query', 'data-plane', 'lsi'] }, () => {
     )
     expect(baseResult.Items).toHaveLength(1)
 
-    await cleanupItems(compositeTableDef.name, [
+    await cleanupItems(compositeIndexedTableDef.name, [
       { pk: noLsiSkItem.pk, sk: noLsiSkItem.sk },
     ])
   })
@@ -272,7 +272,7 @@ describe('Query — LSI pagination across tied sort keys', { tags: ['query', 'da
     await Promise.all(
       tied.map((item) =>
         ddb.send(
-          new PutItemCommand({ TableName: compositeTableDef.name, Item: item }),
+          new PutItemCommand({ TableName: compositeIndexedTableDef.name, Item: item }),
         ),
       ),
     )
@@ -280,7 +280,7 @@ describe('Query — LSI pagination across tied sort keys', { tags: ['query', 'da
 
   afterAll(async () => {
     await cleanupItems(
-      compositeTableDef.name,
+      compositeIndexedTableDef.name,
       tied.map((item) => ({ pk: item.pk, sk: item.sk })),
     )
   })
@@ -293,7 +293,7 @@ describe('Query — LSI pagination across tied sort keys', { tags: ['query', 'da
     do {
       const page = await ddb.send(
         new QueryCommand({
-          TableName: compositeTableDef.name,
+          TableName: compositeIndexedTableDef.name,
           IndexName: 'lsi1',
           KeyConditionExpression: 'pk = :p AND lsi1sk = :v',
           ExpressionAttributeValues: {
