+++ tests/tier3/error-messages/putItem.test.ts
@@ -6,12 +6,11 @@ import { ddb } from '../../../src/client.js'
 import {
   hashTableDef,
   hashBTableDef,
-  gsiBTableDef,
   cleanupItems,
   declareTables,
 } from '../../../src/helpers.js'
 
-declareTables(hashTableDef, hashBTableDef, gsiBTableDef)
+declareTables(hashTableDef, hashBTableDef)
 
 const keysToCleanup = [
   { pk: { S: 'em-put-null-false' } },
@@ -371,26 +370,6 @@ describe('PutItem — exact error messages', { tags: ['put-item', 'data-plane'] 
       expect((err as DynamoDBServiceException).name).toBe('ValidationException')
       expect((err as DynamoDBServiceException).message).toBe(
         'One or more parameter values are not valid. The AttributeValue for a key attribute cannot contain an empty binary value. Key: pk',
-      )
-    }
-  })
-
-  it('empty-binary index key value: full secondary-index-key message', async () => {
-    // Real AWS rejects a zero-length binary value on a secondary index key with
-    // the put-form secondary-index message, naming the index and key.
-    try {
-      await ddb.send(
-        new PutItemCommand({
-          TableName: gsiBTableDef.name,
-          Item: { pk: { S: 'eb-idx' }, bidx: { B: new Uint8Array([]) } },
-        }),
-      )
-      expect.unreachable('should have thrown')
-    } catch (err) {
-      expect(err).toBeInstanceOf(DynamoDBServiceException)
-      expect((err as DynamoDBServiceException).name).toBe('ValidationException')
-      expect((err as DynamoDBServiceException).message).toBe(
-        'One or more parameter values are not valid. A value specified for a secondary index key is not supported. The AttributeValue for a key attribute cannot contain an empty binary value. IndexName: gsib, IndexKey: bidx',
       )
     }
   })
