+++ tests/tier3/error-messages/transactWriteItems.test.ts
@@ -10,13 +10,11 @@ import { skipUnlessSupported } from '../../../src/infra.js'
 import {
   hashTableDef,
   hashBTableDef,
-  gsiBTableDef,
-  compositeTableDef,
   cleanupItems,
   declareTables,
 } from '../../../src/helpers.js'
 
-declareTables(hashTableDef, hashBTableDef, gsiBTableDef, compositeTableDef)
+declareTables(hashTableDef, hashBTableDef)
 
 const keysToCleanup = [
   { pk: { S: 'em-twi-dup' } },
@@ -26,20 +24,8 @@ const keysToCleanup = [
   { pk: { S: 'em-twi-pos-new' } },
 ]
 
-// Defensive: the invalid-key-value cases below fail validation and write
-// nothing; clean up anyway in case a too-lenient target persists them.
-const compositeKeysToCleanup = [
-  { pk: { S: 'em-twi-idx-type' }, sk: { S: 'a' } },
-  { pk: { S: 'em-twi-idx-nonscalar' }, sk: { S: 'b' } },
-  { pk: { S: 'em-twi-idx-empty' }, sk: { S: 'c' } },
-  { pk: { S: 'em-twi-idx-upd-type' }, sk: { S: 'a' } },
-  { pk: { S: 'em-twi-idx-upd-nonscalar' }, sk: { S: 'b' } },
-  { pk: { S: 'em-twi-idx-upd-empty' }, sk: { S: 'c' } },
-]
-
 afterAll(async () => {
   await cleanupItems(hashTableDef.name, keysToCleanup)
-  await cleanupItems(compositeTableDef.name, compositeKeysToCleanup)
 })
 
 describe('TransactWriteItems — exact error messages', { tags: ['transactions', 'data-plane', 'negative-path'] }, () => {
@@ -250,12 +236,12 @@ describe('TransactWriteItems — exact error messages', { tags: ['transactions',
     }
   })
 
-  // Invalid key value (table or index) inside a transact Put/Update. The error
-  // shape splits by fault, captured from real AWS eu-west-2:
+  // Invalid table key value inside a transact Put/Update. The error shape
+  // splits by fault, captured from real AWS eu-west-2:
   //   - wrong type / non-scalar: TransactionCanceledException, reason code
   //     'ValidationError', reason Message carrying the PutItem-style string;
   //   - empty string: top-level ValidationException (up-front input validation).
-  // Index-key messages name gsi1, the alphabetically-first index lsi1sk keys.
+  // The secondary-index equivalents live in transactIndexKeys.test.ts.
   const expectCancelledReason = async (
     command: TransactWriteItemsCommand,
     reasonMessage: string,
@@ -309,114 +295,12 @@ describe('TransactWriteItems — exact error messages', { tags: ['transactions',
     )
   })
 
-  it('Put wrong-typed index key: cancelled with full ValidationError reason', async () => {
-    await expectCancelledReason(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Put: {
-              TableName: compositeTableDef.name,
-              Item: { pk: { S: 'em-twi-idx-type' }, sk: { S: 'a' }, lsi1sk: { N: '5' } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values were invalid: Type mismatch for Index Key lsi1sk Expected: S Actual: N IndexName: gsi1',
-    )
-  })
-
-  it('Put non-scalar index key: cancelled with full ValidationError reason', async () => {
-    await expectCancelledReason(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Put: {
-              TableName: compositeTableDef.name,
-              Item: { pk: { S: 'em-twi-idx-nonscalar' }, sk: { S: 'b' }, lsi1sk: { L: [{ S: 'x' }] } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values were invalid: Type mismatch for Index Key lsi1sk Expected: S Actual: L IndexName: gsi1',
-    )
-  })
-
-  it('Update wrong-typed index key: cancelled with full ValidationError reason', async () => {
-    await expectCancelledReason(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Update: {
-              TableName: compositeTableDef.name,
-              Key: { pk: { S: 'em-twi-idx-upd-type' }, sk: { S: 'a' } },
-              UpdateExpression: 'SET lsi1sk = :v',
-              ExpressionAttributeValues: { ':v': { N: '5' } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values were invalid: Type mismatch for Index Key lsi1sk Expected: S Actual: N IndexName: gsi1',
-    )
-  })
-
-  it('Update non-scalar index key: cancelled with full ValidationError reason', async () => {
-    await expectCancelledReason(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Update: {
-              TableName: compositeTableDef.name,
-              Key: { pk: { S: 'em-twi-idx-upd-nonscalar' }, sk: { S: 'b' } },
-              UpdateExpression: 'SET lsi1sk = :v',
-              ExpressionAttributeValues: { ':v': { L: [{ S: 'x' }] } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values were invalid: Type mismatch for Index Key lsi1sk Expected: S Actual: L IndexName: gsi1',
-    )
-  })
-
   it('Put empty-string table key: top-level ValidationException', async () => {
     await expectTopLevelValidation(
       new TransactWriteItemsCommand({
         TransactItems: [{ Put: { TableName: hashTableDef.name, Item: { pk: { S: '' } } } }],
       }),
       'One or more parameter values are not valid. The AttributeValue for a key attribute cannot contain an empty string value. Key: pk',
-    )
-  })
-
-  it('Put empty-string index key: top-level ValidationException', async () => {
-    await expectTopLevelValidation(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Put: {
-              TableName: compositeTableDef.name,
-              Item: { pk: { S: 'em-twi-idx-empty' }, sk: { S: 'c' }, lsi1sk: { S: '' } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values are not valid. A value specified for a secondary index key is not supported. The AttributeValue for a key attribute cannot contain an empty string value. IndexName: gsi1, IndexKey: lsi1sk',
-    )
-  })
-
-  it('Update empty-string index key: top-level ValidationException', async () => {
-    await expectTopLevelValidation(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Update: {
-              TableName: compositeTableDef.name,
-              Key: { pk: { S: 'em-twi-idx-upd-empty' }, sk: { S: 'c' } },
-              UpdateExpression: 'SET lsi1sk = :v',
-              ExpressionAttributeValues: { ':v': { S: '' } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values are not valid. The update expression attempted to update a secondary index key to a value that is not supported. The AttributeValue for a key attribute cannot contain an empty string value.',
     )
   })
 
@@ -469,33 +353,4 @@ describe('TransactWriteItems — exact error messages', { tags: ['transactions',
     expectTopLevelValidation(cmd(delKeyB(emptyBin)), emptyBinKeyMsg))
   it('ConditionCheck empty-binary Key: top-level empty-value message', () =>
     expectTopLevelValidation(cmd(ccKeyB(emptyBin)), emptyBinKeyMsg))
-
-  it('Put empty-binary index key: top-level ValidationException', async () => {
-    await expectTopLevelValidation(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          { Put: { TableName: gsiBTableDef.name, Item: { pk: { S: 'eb-twi-idx' }, bidx: { B: new Uint8Array([]) } } } },
-        ],
-      }),
-      'One or more parameter values are not valid. A value specified for a secondary index key is not supported. The AttributeValue for a key attribute cannot contain an empty binary value. IndexName: gsib, IndexKey: bidx',
-    )
-  })
-
-  it('Update empty-binary index key: top-level ValidationException', async () => {
-    await expectTopLevelValidation(
-      new TransactWriteItemsCommand({
-        TransactItems: [
-          {
-            Update: {
-              TableName: gsiBTableDef.name,
-              Key: { pk: { S: 'eb-twi-idx-upd' } },
-              UpdateExpression: 'SET bidx = :v',
-              ExpressionAttributeValues: { ':v': { B: new Uint8Array([]) } },
-            },
-          },
-        ],
-      }),
-      'One or more parameter values are not valid. The update expression attempted to update a secondary index key to a value that is not supported. The AttributeValue for a key attribute cannot contain an empty binary value.',
-    )
-  })
 })
